+++ src/redgrease/sugar.py
@@ -27,6 +27,44 @@ class TriggerMode:
     Sync = "sync"
 
 
+class KeyType:
+    """Redis Key Types"""
+
+    String = "string"
+    Hash = "hash"
+    List = "list"
+    Set = "set"
+    ZSet = "zset"
+    Stream = "stream"
+    Module = "module"
+
+    _constructors = {
+        "string": str,
+        "hash": dict,
+        "list": list,
+        "set": set,
+    }
+
+    @staticmethod
+    def of(python_type):
+        if python_type is str or isinstance(python_type, str):
+            return KeyType.String
+        elif python_type is dict or isinstance(python_type, dict):
+            return KeyType.Hash
+        elif python_type is list or isinstance(python_type, list):
+            return KeyType.List
+        elif python_type is set or isinstance(python_type, set):
+            return KeyType.Set
+        else:
+            raise ValueError(
+                f"No corresponding Redis Key-Type for python type {python_type}"
+            )
+
+    @staticmethod
+    def constructor(key_type: str):
+        return KeyType._constructors[key_type]
+
+
 class LogLevel:
     """Redis Gears log levels"""
